+++ packages/orq-rc/RELEASES.md
@@ -1529,3 +1529,13 @@ Based on:
 - [typescript v3.6.0-rc.9] packages/orq-rc
 ### Releases
 - [NPM v3.6.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/3.6.0-rc.9 - packages/orq-rc
+
+## 2025-05-18 21:01:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.546.2 (2.604.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.6.0-rc.10] packages/orq-rc
+### Releases
+- [NPM v3.6.0-rc.10] https://www.npmjs.com/package/@orq-ai/node/v/3.6.0-rc.10 - packages/orq-rc